util/requirements.txt
python-can
intelhex
